% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{as.formant}
\alias{as.formant}
\title{as.formant}
\usage{
as.formant(formant, name = "")
}
\arguments{
\item{formant}{Formant 2 object}

\item{name}{New name}
}
\value{
Formant 2 object
}
\description{
Renames the class(formant)["name"] attribute and sets class(formant)["type"] <- "Formant 2" (if it is not already set)
}
\examples{
class(formant.sample())
class(as.formant(formant.sample(), name = "New Name"))
}
