% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it.R
\name{as.it}
\alias{as.it}
\title{as.it}
\usage{
as.it(it, name = "")
}
\arguments{
\item{it}{IntensityTier object}

\item{name}{New name}
}
\value{
IntensityTier object
}
\description{
Renames the \code{class(it)["name"]} attribute and sets \code{class(it)["type"] <- "IntensityTier"} (if it is not already set)
}
\examples{
class(it.sample())
class(as.it(it.sample(), name = "New Name"))
}
