% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{as.pitch}
\alias{as.pitch}
\title{as.pitch}
\usage{
as.pitch(pitch, name = "")
}
\arguments{
\item{pitch}{Pitch 1 object}

\item{name}{New name}
}
\value{
Pitch 1 object
}
\description{
Renames the class(pitch)["name"] attribute and sets class(pitch)["type"] <- "Pitch 1" (if it is not already set)
}
\examples{
class(pitch.sample())
class(as.pitch(pitch.sample(), name = "New Name"))
}
