% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{as.pt}
\alias{as.pt}
\title{as.pt}
\usage{
as.pt(pt, name = "")
}
\arguments{
\item{pt}{PitchTier object}

\item{name}{New name}
}
\value{
PitchTier object
}
\description{
Renames the \code{class(pt)["name"]} attribute and sets \code{class(pt)["type"] <- "PitchTier"} (if it is not already set)
}
\examples{
class(pt.sample())
class(as.pt(pt.sample(), name = "New Name"))
}
