% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snd.R
\name{as.snd}
\alias{as.snd}
\title{as.snd}
\usage{
as.snd(snd, name = "")
}
\arguments{
\item{snd}{snd object}

\item{name}{New name}
}
\value{
snd object
}
\description{
Renames the \code{class(snd)["name"]} attribute and sets \code{class(snd)["type"] <- "Sound"} (if it is not already set)
}
\examples{
class(snd.sample())
class(as.snd(snd.sample(), name = "New Name"))
}
