% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{as.sound}
\alias{as.sound}
\title{as.sound}
\usage{
as.sound(snd, name = "")
}
\arguments{
\item{snd}{snd object}

\item{name}{New name}
}
\value{
snd object
}
\description{
Renames the class(snd)["name"] attribute and sets class(snd)["type"] <- "Sound" (if it is not already set)
}
\examples{
class(snd.sample())
class(as.snd(snd.sample(), name = "New Name"))
}
