% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg.R
\name{as.tg}
\alias{as.tg}
\title{as.tg}
\usage{
as.tg(tg, name = "")
}
\arguments{
\item{tg}{TextGrid object}

\item{name}{New name}
}
\value{
TextGrid object
}
\description{
Renames the \code{class(tg)["name"]} attribute and sets {class(tg)["type"] <- "TextGrid"} (if it is not already set)
}
\examples{
class(tg.sample())
class(as.tg(tg.sample(), name = "New Name"))
}
