% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{col.read}
\alias{col.read}
\title{col.read}
\usage{
col.read(fileName, encoding = "UTF-8")
}
\arguments{
\item{fileName}{Input file name}

\item{encoding}{File encoding (default: "UTF-8"), "auto" for auto-detect of Unicode encoding}
}
\value{
Collection object
}
\description{
Loads Collection from Praat in Text or Short text format.
Collection may contain combination of TextGrids, PitchTiers, Pitch objects, and IntensityTiers.
}
\examples{
\dontrun{
coll <- col.read("short_textgrid+pitchtier.Collection")
length(coll)  # number of objects in collection
class(coll[[1]])["type"]  # 1st object type
class(coll[[1]])["name"]  # 1st object name
tg1 <- coll[[1]]  # 1st object
tg.plot(tg1)
length(tg1)  # number of tiers in TextGrid
tg1$word$label

class(coll[[2]])["type"]  # 2nd object type
class(coll[[2]])["name"]  # 2nd object type
pt1 <- coll[[2]]  # 2nd object
pt.plot(pt1)
}
}
\seealso{
\code{\link{tg.read}}, \code{\link{pt.read}}, \code{\link{pitch.read}}, \code{\link{it.read}}
}
