% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{detectEncoding}
\alias{detectEncoding}
\title{detectEncoding}
\usage{
detectEncoding(fileName)
}
\arguments{
\item{fileName}{Input file name}
}
\value{
detected encoding of the text input file
}
\description{
Detects unicode encoding of Praat text files
}
\examples{
\dontrun{
detectEncoding("demo/H.TextGrid")
detectEncoding("demo/H_UTF16.TextGrid")
}
}
