% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formant.R
\name{formant.getPointIndexHigherThanTime}
\alias{formant.getPointIndexHigherThanTime}
\title{formant.getPointIndexHigherThanTime}
\usage{
formant.getPointIndexHigherThanTime(formant, time)
}
\arguments{
\item{formant}{Formant object}

\item{time}{time which is going to be found in frames}
}
\value{
integer
}
\description{
Returns index of frame which is nearest the given time from right, i.e.
\code{time} <= frameTime.
}
\examples{
formant <- formant.sample()
formant.getPointIndexHigherThanTime(formant, 0.5)
}
\seealso{
\code{\link{formant.getPointIndexNearestTime}}, \code{\link{formant.getPointIndexLowerThanTime}}
}
