% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formant.R
\name{formant.getPointIndexLowerThanTime}
\alias{formant.getPointIndexLowerThanTime}
\title{formant.getPointIndexLowerThanTime}
\usage{
formant.getPointIndexLowerThanTime(formant, time)
}
\arguments{
\item{formant}{Formant object}

\item{time}{time which is going to be found in frames}
}
\value{
integer
}
\description{
Returns index of frame which is nearest the given time from left, i.e.
frameTime <= \code{time}.
}
\examples{
formant <- formant.sample()
formant.getPointIndexLowerThanTime(formant, 0.5)
}
\seealso{
\code{\link{formant.getPointIndexNearestTime}}, \code{\link{formant.getPointIndexHigherThanTime}}
}
