% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formant.R
\name{formant.getPointIndexNearestTime}
\alias{formant.getPointIndexNearestTime}
\title{formant.getPointIndexNearestTime}
\usage{
formant.getPointIndexNearestTime(formant, time)
}
\arguments{
\item{formant}{Formant object}

\item{time}{time which is going to be found in frames}
}
\value{
integer
}
\description{
Returns index of frame which is nearest the given \code{time} (from both sides).
}
\examples{
formant <- formant.sample()
formant.getPointIndexNearestTime(formant, 0.5)
}
\seealso{
\code{\link{formant.getPointIndexLowerThanTime}}, \code{\link{formant.getPointIndexHigherThanTime}}
}
