% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{formant.plot}
\alias{formant.plot}
\title{formant.plot}
\usage{
formant.plot(formant, scaleIntensity = TRUE, drawBandwidth = FALSE,
  group = "")
}
\arguments{
\item{formant}{Formant object}

\item{scaleIntensity}{Point size scaled according to relative intensity}

\item{drawBandwidth}{Draw formant bandwidth}

\item{group}{[optional] character string, name of group for dygraphs synchronization}
}
\description{
Plots interactive Formant object using dygraphs package.
}
\examples{
\dontrun{
formant <- formant.sample()
formant.plot(formant, drawBandwidth = TRUE)
}
}
\seealso{
\code{\link{formant.read}}, \code{\link{formant.sample}}, \code{\link{formant.toArray}}, \code{\link{tg.plot}}
}
