% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat_sampleData.R
\name{formant.sample}
\alias{formant.sample}
\title{formant.sample}
\usage{
formant.sample()
}
\value{
Formant
}
\description{
Returns sample Formant object.
}
\examples{
formant <- formant.sample()
}
\seealso{
\code{\link{tg.sample}}, \code{\link{pt.sample}}, \code{\link{it.sample}}, \code{\link{pitch.sample}}
}
