% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{formant.toArray}
\alias{formant.toArray}
\title{formant.toArray}
\usage{
formant.toArray(formant)
}
\arguments{
\item{formant}{Formant object}
}
\value{
Formant object with frames converted to frequency and bandwidth arrays and intensity vector
}
\description{
formant.toArray
}
\examples{
formantArray <- formant.toArray(formant.sample())
formantArray$t[1:10]
formantArray$frequencyArray[, 1:10]
formantArray$bandwidthArray[, 1:10]
formantArray$intensityVector[1:10]
\dontrun{
plot(formantArray$t, formantArray$frequencyArray[1, ]) # draw 1st formant track
}
}
\seealso{
\code{\link{formant.read}}, \code{\link{formant.plot}}
}
