% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{formant.toFrame}
\alias{formant.toFrame}
\title{formant.toFrame}
\usage{
formant.toFrame(formantArray)
}
\arguments{
\item{formantArray}{Formant object (array format)}
}
\value{
Formant object with frames
}
\description{
formant.toFrame
}
\examples{
formantArray <- formant.toArray(formant.sample())
formant <- formant.toFrame(formantArray)
}
\seealso{
\code{\link{formant.toArray}}, \code{\link{formant.read}}, \code{\link{formant.plot}}
}
