% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formant.R
\name{formant.write}
\alias{formant.write}
\title{formant.write}
\usage{
formant.write(formant, fileNameFormant, format = "short")
}
\arguments{
\item{formant}{Formant object}

\item{fileNameFormant}{Output file name}

\item{format}{Output file format (\code{"short"} (default, short text format) or \code{"text"} (a.k.a. full text format))}
}
\description{
Saves Formant to the file.
}
\examples{
\dontrun{
formant <- formant.sample()
formant.write(formant, "demo_output.Formant")
}
}
\seealso{
\code{\link{formant.read}}, \code{\link{tg.read}}
}
