% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{isInt}
\alias{isInt}
\title{isInt}
\usage{
isInt(num)
}
\arguments{
\item{num}{variable to be tested}
}
\value{
TRUE / FALSE
}
\description{
Returns TRUE / FALSE whether it is exactly 1 integer number (in fact, the class can be numeric but the number must be integer), non-missing
}
\examples{
isInt(2)
isInt(2L)
isInt(-2)
isInt(-2L)
isInt(2.1)
isInt(-2.1)
isInt(1:5)
isInt(NA_integer_)
isInt(integer(0))
}
\seealso{
\code{\link{isNum}}, \code{\link{isLogical}}, \code{\link{isString}}
}
