% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{isLogical}
\alias{isLogical}
\title{isLogical}
\usage{
isLogical(logical)
}
\arguments{
\item{logical}{variable to be tested}
}
\value{
\code{TRUE} / \code{FALSE}
}
\description{
Returns \code{TRUE} / \code{FALSE} whether it is exactly 1 logical value, non-missing
}
\examples{
isLogical(TRUE)
isLogical(FALSE)
isLogical(1)
isLogical(0)
isLogical(2)
isLogical(NA)
isLogical(NaN)
isLogical(logical(0))
}
\seealso{
\code{\link{isNum}}, \code{\link{isInt}}, \code{\link{isString}}
}
