% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{isNum}
\alias{isNum}
\title{isNum}
\usage{
isNum(num)
}
\arguments{
\item{num}{variable to be tested}
}
\value{
TRUE / FALSE
}
\description{
Returns TRUE / FALSE whether it is exactly 1 number (numeric or integer vector of length 1, non-missing)
}
\examples{
isNum(2)
isNum(2L)
isNum(-2)
isNum(-2L)
isNum(2.1)
isNum(-2.1)
isNum(1:5)
isNum(NA_real_)
isNum(numeric(0))
}
\seealso{
\code{\link{isInt}}, \code{\link{isLogical}}, \code{\link{isString}}
}
