% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{isString}
\alias{isString}
\title{isString}
\usage{
isString(string)
}
\arguments{
\item{string}{variable to be tested}
}
\value{
\code{TRUE} / \code{FALSE}
}
\description{
Returns \code{TRUE} / \code{FALSE} whether it is exactly 1 character string (character vector of length 1, non-missing)
}
\examples{
isString("hello")
isString(2)
isString(c("hello", "world"))
isString(NA_character_)
}
\seealso{
\code{\link{isInt}}, \code{\link{isNum}}, \code{\link{isLogical}}
}
