% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it.R
\name{it.getPointIndexHigherThanTime}
\alias{it.getPointIndexHigherThanTime}
\title{it.getPointIndexHigherThanTime}
\usage{
it.getPointIndexHigherThanTime(it, time)
}
\arguments{
\item{it}{IntensityTier object}

\item{time}{time which is going to be found in points}
}
\value{
integer
}
\description{
Returns index of point which is nearest the given time from right, i.e.
\code{time} <= pointTime.
}
\examples{
it <- it.sample()
it.getPointIndexHigherThanTime(it, 0.5)
}
\seealso{
\code{\link{it.getPointIndexNearestTime}}, \code{\link{it.getPointIndexLowerThanTime}}
}
