% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it.R
\name{it.getPointIndexLowerThanTime}
\alias{it.getPointIndexLowerThanTime}
\title{it.getPointIndexLowerThanTime}
\usage{
it.getPointIndexLowerThanTime(it, time)
}
\arguments{
\item{it}{IntensityTier object}

\item{time}{time which is going to be found in points}
}
\value{
integer
}
\description{
Returns index of point which is nearest the given time from left, i.e.
pointTime <= \code{time}.
}
\examples{
it <- it.sample()
it.getPointIndexLowerThanTime(it, 0.5)
}
\seealso{
\code{\link{it.getPointIndexNearestTime}}, \code{\link{it.getPointIndexHigherThanTime}}
}
