% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it.R
\name{it.getPointIndexNearestTime}
\alias{it.getPointIndexNearestTime}
\title{it.getPointIndexNearestTime}
\usage{
it.getPointIndexNearestTime(it, time)
}
\arguments{
\item{it}{IntensityTier object}

\item{time}{time which is going to be found in points}
}
\value{
integer
}
\description{
Returns index of point which is nearest the given \code{time} (from both sides).
}
\examples{
it <- it.sample()
it.getPointIndexNearestTime(it, 0.5)
}
\seealso{
\code{\link{it.getPointIndexLowerThanTime}}, \code{\link{it.getPointIndexHigherThanTime}}
}
