% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it.R
\name{it.legendreDemo}
\alias{it.legendreDemo}
\title{it.legendreDemo}
\usage{
it.legendreDemo()
}
\description{
Plots first four Legendre polynomials
}
\examples{
\dontrun{
it.legendreDemo()
}
}
\seealso{
\code{\link{it.legendre}}, \code{\link{it.legendreSynth}}, \code{\link{it.read}}, \code{\link{it.plot}}, \code{\link{it.interpolate}}
}
