% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{it.read}
\alias{it.read}
\title{it.read}
\usage{
it.read(fileNameIntensityTier, encoding = "UTF-8")
}
\arguments{
\item{fileNameIntensityTier}{file name of IntensityTier}

\item{encoding}{File encoding (default: "UTF-8"), "auto" for auto-detect of Unicode encoding}
}
\value{
IntensityTier object
}
\description{
Reads IntensityTier from Praat. Supported formats: text file, short text file.
}
\examples{
\dontrun{
it <- it.read("demo/maminka.IntensityTier")
it.plot(it)
}
}
\seealso{
\code{\link{it.write}}, \code{\link{it.plot}}, \code{\link{it.cut}}, \code{\link{it.cut0}}, \code{\link{it.interpolate}}, \code{\link{tg.read}}, \code{\link{pt.read}}, \code{\link{pitch.read}}, \code{\link{col.read}}
}
