% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat_sampleData.R
\name{it.sample}
\alias{it.sample}
\title{it.sample}
\usage{
it.sample()
}
\value{
IntensityTier
}
\description{
Returns sample IntensityTier.
}
\examples{
it <- it.sample()
it.plot(it)
}
\seealso{
\code{\link{it.plot}}
}
