% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{pitch.getPointIndexLowerThanTime}
\alias{pitch.getPointIndexLowerThanTime}
\title{pitch.getPointIndexLowerThanTime}
\usage{
pitch.getPointIndexLowerThanTime(pitch, time)
}
\arguments{
\item{pitch}{Pitch object}

\item{time}{time which is going to be found in frames}
}
\value{
integer
}
\description{
Returns index of frame which is nearest the given time from left, i.e.
frameTime <= \code{time}.
}
\examples{
pitch <- pitch.sample()
pitch.getPointIndexLowerThanTime(pitch, 0.5)
}
\seealso{
\code{\link{pitch.getPointIndexNearestTime}}, \code{\link{pitch.getPointIndexHigherThanTime}}
}
