% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{pitch.getPointIndexNearestTime}
\alias{pitch.getPointIndexNearestTime}
\title{pitch.getPointIndexNearestTime}
\usage{
pitch.getPointIndexNearestTime(pitch, time)
}
\arguments{
\item{pitch}{Pitch object}

\item{time}{time which is going to be found in frames}
}
\value{
integer
}
\description{
Returns index of frame which is nearest the given \code{time} (from both sides).
}
\examples{
pitch <- pitch.sample()
pitch.getPointIndexNearestTime(pitch, 0.5)
}
\seealso{
\code{\link{pitch.getPointIndexLowerThanTime}}, \code{\link{pitch.getPointIndexHigherThanTime}}
}
