% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{pitch.plot}
\alias{pitch.plot}
\title{pitch.plot}
\usage{
pitch.plot(pitch, scaleIntensity = TRUE, showStrength = TRUE,
  group = "")
}
\arguments{
\item{pitch}{Pitch object}

\item{scaleIntensity}{Point size scaled according to relative intensity}

\item{showStrength}{show strength annotation (TRUE/FALSE)}

\item{group}{[optional] character string, name of group for dygraphs synchronization}
}
\description{
Plots interactive Pitch object using dygraphs package.
}
\examples{
\dontrun{
pitch <- pitch.sample()
pitch.plot(pitch)
}
}
\seealso{
\code{\link{pitch.read}}, \code{\link{pitch.sample}}, \code{\link{pitch.toArray}}, \code{\link{tg.plot}}
}
