% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat_sampleData.R
\name{pitch.sample}
\alias{pitch.sample}
\title{pitch.sample}
\usage{
pitch.sample()
}
\value{
Pitch
}
\description{
Returns sample Pitch object.
}
\examples{
pitch <- pitch.sample()
}
\seealso{
\code{\link{tg.sample}}, \code{\link{pt.sample}}, \code{\link{it.sample}}, \code{\link{formant.sample}}
}
