% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{pitch.toArray}
\alias{pitch.toArray}
\title{pitch.toArray}
\usage{
pitch.toArray(pitch)
}
\arguments{
\item{pitch}{Pitch object (frame format)}
}
\value{
Pitch object with frames converted to frequency and strength arrays and intensity vector
}
\description{
pitch.toArray
}
\examples{
pitchArray <- pitch.toArray(pitch.sample())
pitchArray$t[1:10]
pitchArray$frequencyArray[, 1:10]
pitchArray$bandwidthArray[, 1:10]
pitchArray$intensityVector[1:10]
}
\seealso{
\code{\link{pitch.toFrame}}, \code{\link{pitch.read}}, \code{\link{pitch.plot}}
}
