% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{pitch.toFrame}
\alias{pitch.toFrame}
\title{pitch.toFrame}
\usage{
pitch.toFrame(pitchArray)
}
\arguments{
\item{pitchArray}{Pitch object (array format)}
}
\value{
Pitch object with frames
}
\description{
pitch.toFrame
}
\examples{
pitchArray <- pitch.toArray(pitch.sample())
pitch <- pitch.toFrame(pitchArray)
}
\seealso{
\code{\link{pitch.toArray}}, \code{\link{pitch.read}}, \code{\link{pitch.plot}}
}
