% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{pitch.write}
\alias{pitch.write}
\title{pitch.write}
\usage{
pitch.write(pitch, fileNamePitch, format = "short")
}
\arguments{
\item{pitch}{Pitch object}

\item{fileNamePitch}{Output file name}

\item{format}{Output file format (\code{"short"} (default, short text format) or \code{"text"} (a.k.a. full text format))}
}
\description{
Saves Pitch to the file.
}
\examples{
\dontrun{
pitch <- pitch.sample()
pitch.write(pitch, "demo_output.Pitch")
}
}
\seealso{
\code{\link{pitch.read}}, \code{\link{pt.read}}
}
