% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pt.getPointIndexHigherThanTime}
\alias{pt.getPointIndexHigherThanTime}
\title{pt.getPointIndexHigherThanTime}
\usage{
pt.getPointIndexHigherThanTime(pt, time)
}
\arguments{
\item{pt}{PitchTier object}

\item{time}{time which is going to be found in points}
}
\value{
integer
}
\description{
Returns index of point which is nearest the given time from right, i.e.
\code{time} <= pointTime.
}
\examples{
pt <- pt.sample()
pt.getPointIndexHigherThanTime(pt, 0.5)
}
\seealso{
\code{\link{pt.getPointIndexNearestTime}}, \code{\link{pt.getPointIndexLowerThanTime}}
}
