% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pt.getPointIndexLowerThanTime}
\alias{pt.getPointIndexLowerThanTime}
\title{pt.getPointIndexLowerThanTime}
\usage{
pt.getPointIndexLowerThanTime(pt, time)
}
\arguments{
\item{pt}{PitchTier object}

\item{time}{time which is going to be found in points}
}
\value{
integer
}
\description{
Returns index of point which is nearest the given time from left, i.e.
pointTime <= \code{time}.
}
\examples{
pt <- pt.sample()
pt.getPointIndexLowerThanTime(pt, 0.5)
}
\seealso{
\code{\link{pt.getPointIndexNearestTime}}, \code{\link{pt.getPointIndexHigherThanTime}}
}
