% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pt.getPointIndexNearestTime}
\alias{pt.getPointIndexNearestTime}
\title{pt.getPointIndexNearestTime}
\usage{
pt.getPointIndexNearestTime(pt, time)
}
\arguments{
\item{pt}{PitchTier object}

\item{time}{time which is going to be found in points}
}
\value{
integer
}
\description{
Returns index of point which is nearest the given \code{time} (from both sides).
}
\examples{
pt <- pt.sample()
pt.getPointIndexNearestTime(pt, 0.5)
}
\seealso{
\code{\link{pt.getPointIndexLowerThanTime}}, \code{\link{pt.getPointIndexHigherThanTime}}
}
