% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{pt.legendreDemo}
\alias{pt.legendreDemo}
\title{pt.legendreDemo}
\usage{
pt.legendreDemo()
}
\description{
Plots first four Legendre polynomials
}
\examples{
pt.legendreDemo()

}
\seealso{
\code{\link{pt.legendre}}, \code{\link{pt.legendreSynth}}, \code{\link{pt.read}}, \code{\link{pt.plot}}, \code{\link{pt.Hz2ST}}, \code{\link{pt.interpolate}}
}
