% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{pt.legendreSynth}
\alias{pt.legendreSynth}
\title{pt.legendreSynth}
\usage{
pt.legendreSynth(c, npoints = 1000)
}
\arguments{
\item{c}{Vector of Legendre polynomials}

\item{npoints}{Number of points of PitchTier interpolation}
}
\value{
Vector of values of synthetized contour
}
\description{
Synthetize the contour from vector of Legendre polynomials 'c' in 'npoints' equidistant points
}
\examples{
pt <- pt.sample()
pt$f <- pt$f[pt$t >= 3]  # Cut the PitchTier - from time 3 seconds to the end
pt$t <- pt$t[pt$t >= 3] - 3
pt$tmax <- pt$tmax - 3
c <- pt.legendre(pt)
print(c)
leg <- pt.legendreSynth(c)
\dontrun{
pt.plot(pt)
plot(leg)
}
}
\seealso{
\code{\link{pt.legendre}}, \code{\link{pt.legendreDemo}}, \code{\link{pt.read}}, \code{\link{pt.plot}}, \code{\link{pt.Hz2ST}}, \code{\link{pt.interpolate}}
}
