% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{pt.plot}
\alias{pt.plot}
\title{pt.plot}
\usage{
pt.plot(pt, group = "")
}
\arguments{
\item{pt}{PitchTier object}

\item{group}{[optional] character string, name of group for dygraphs synchronization}
}
\description{
Plots interactive PitchTier using dygraphs package.
}
\examples{
\dontrun{
pt <- pt.sample()
pt.plot(pt)
}
}
\seealso{
\code{\link{pt.read}}, \code{\link{tg.plot}}, \code{\link{pt.Hz2ST}}, \code{\link{pt.interpolate}}, \code{\link{pt.write}}
}
