% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snd.R
\name{snd.getPointIndexHigherThanTime}
\alias{snd.getPointIndexHigherThanTime}
\title{snd.getPointIndexHigherThanTime}
\usage{
snd.getPointIndexHigherThanTime(snd, time)
}
\arguments{
\item{snd}{Sound object}

\item{time}{time which is going to be found in samples}
}
\value{
integer
}
\description{
Returns index of sample which is nearest the given time from right, i.e.
\code{time} <= sampleTime.
}
\examples{
snd <- snd.sample()
snd.getPointIndexHigherThanTime(snd, 0.5)
}
\seealso{
\code{\link{snd.getPointIndexNearestTime}}, \code{\link{snd.getPointIndexLowerThanTime}}
}
