% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snd.R
\name{snd.getPointIndexNearestTime}
\alias{snd.getPointIndexNearestTime}
\title{snd.getPointIndexNearestTime}
\usage{
snd.getPointIndexNearestTime(snd, time)
}
\arguments{
\item{snd}{Sound object}

\item{time}{time which is going to be found in samples}
}
\value{
integer
}
\description{
Returns index of sample which is nearest the given \code{time} (from both sides).
}
\examples{
snd <- snd.sample()
snd.getPointIndexNearestTime(snd, 0.5)
}
\seealso{
\code{\link{snd.getPointIndexLowerThanTime}}, \code{\link{snd.getPointIndexHigherThanTime}}
}
