% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{snd.plot}
\alias{snd.plot}
\title{snd.plot}
\usage{
snd.plot(snd, group = "")
}
\arguments{
\item{snd}{Sound object}

\item{group}{[optional] character string, name of group for dygraphs synchronization}
}
\description{
Plots interactive Sound object using dygraphs package.
}
\examples{
\dontrun{
snd <- snd.sample()
snd.plot(snd)
}
}
\seealso{
\code{\link{snd.read}}
}
