% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat_sampleData.R
\name{snd.sample}
\alias{snd.sample}
\title{snd.sample}
\usage{
snd.sample()
}
\value{
snd
}
\description{
Returns sample Sound object.
}
\examples{
snd <- snd.sample()
snd.plot(snd)
}
\seealso{
\code{\link{snd.plot}}
}
