% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{str_find}
\alias{str_find}
\title{str_find}
\usage{
str_find(string, patternNoRegex)
}
\arguments{
\item{string}{string in which we try to find something}

\item{patternNoRegex}{string we want to find, "as it is" - no regular exprressions}
}
\value{
indices of all occurences (1 = 1st character)
}
\description{
Find string in another string (without regular expressions), returns indices of all occurences.
}
\examples{
str_find("Hello, hello, hello world", "ell")   # 2 9 16
str_find("Hello, hello, hello world", "q")     # integer(0)
}
\seealso{
\code{\link{str_find1}}, \code{\link{str_contains}}, \code{\link{isString}}
}
