% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{str_find1}
\alias{str_find1}
\title{str_find1}
\usage{
str_find1(string, patternNoRegex)
}
\arguments{
\item{string}{string in which we try to find something}

\item{patternNoRegex}{string we want to find, "as it is" - no regular exprressions}
}
\value{
index of the first occurence only (\code{1} = 1st character)
}
\description{
Find string in another string (without regular expressions), returns indices of the first occurence only.
}
\examples{
str_find1("Hello, hello, hello world", "ell")   # 2
str_find1("Hello, hello, hello world", "q")     # integer(0)
}
\seealso{
\code{\link{str_find}}, \code{\link{str_contains}}, \code{\link{isString}}
}
