% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{strTrim}
\alias{strTrim}
\title{strTrim}
\usage{
strTrim(string)
}
\arguments{
\item{string}{character string}
}
\value{
returns a character string with removed leading and trailing whitespace characters.
}
\description{
Trim leading and trailing whitespace in character string.
}
\details{
Like \code{str_trim()} in \code{stringr} package or \code{trimws()} in R3.2.0 but way faster.

Source: Hadley Wickham comment at https://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace-in-r
}
\examples{
strTrim("      Hello World!    ")
}
\seealso{
\code{\link{isString}} for testing whether it is 1 character vector, \code{\link{str_contains}} for finding string in string without regexp, \code{\link{str_find}} for all indices without regexp, \code{\link{str_find1}} for the first index withoud regexp.
}
