% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.checkTierInd}
\alias{tg.checkTierInd}
\title{tg.checkTierInd}
\usage{
tg.checkTierInd(tg, tierInd)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{Tier index or "name"}
}
\value{
Tier index
}
\description{
Returns tier index. Input can be either index (number) or tier name (character string).
It performs checks whether the tier exists.
}
\examples{
tg <- tg.sample()
tg.checkTierInd(tg, 4)
tg.checkTierInd(tg, 'word')
}
\seealso{
\code{\link{tg.getTierName}}, \code{\link{tg.isIntervalTier}}, \code{\link{tg.isPointTier}}, \code{\link{tg.plot}}, \code{\link{tg.getNumberOfTiers}}
}

