% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.countLabels}
\alias{tg.countLabels}
\title{tg.countLabels}
\usage{
tg.countLabels(tg, tierInd, label)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{label}{character string: label to be counted}
}
\value{
integer number
}
\description{
Returns number of labels with the specified label.
}
\examples{
tg <- tg.sample()
tg.countLabels(tg, "phone", "a")
}

