% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getEndTime}
\alias{tg.getEndTime}
\title{tg.getEndTime}
\usage{
tg.getEndTime(tg, tierInd = 0)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{[optional] tier index or "name"}
}
\value{
numeric
}
\description{
Returns end time. If tier index is specified, it returns end time
of the tier, if it is not specified, it returns end time of the whole
TextGrid.
}
\examples{
tg <- tg.sample()
tg.getEndTime(tg)
tg.getEndTime(tg, "phone")
}
\seealso{
\code{\link{tg.getStartTime}}, \code{\link{tg.getTotalDuration}}
}

