% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getIntervalEndTime}
\alias{tg.getIntervalEndTime}
\title{tg.getIntervalEndTime}
\usage{
tg.getIntervalEndTime(tg, tierInd, index)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{index}{index of interval}
}
\value{
numeric
}
\description{
Return end time of interval in interval tier.
}
\examples{
tg <- tg.sample()
tg.getIntervalEndTime(tg, "phone", 5)
}

