% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getIntervalIndexAtTime}
\alias{tg.getIntervalIndexAtTime}
\title{tg.getIntervalIndexAtTime}
\usage{
tg.getIntervalIndexAtTime(tg, tierInd, time)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{time}{time which is going to be found in intervals}
}
\value{
integer
}
\description{
Returns index of interval which includes the given time, i.e.
tStart <= time < tEnd. Tier index must belong to interval tier.
}
\examples{
tg <- tg.sample()
tg.getIntervalIndexAtTime(tg, "word", 0.5)
}
\seealso{
\code{\link{tg.getIntervalStartTime}}, \code{\link{tg.getIntervalEndTime}}, \code{\link{tg.getLabel}}
}
