% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getIntervalStartTime}
\alias{tg.getIntervalStartTime}
\title{tg.getIntervalStartTime}
\usage{
tg.getIntervalStartTime(tg, tierInd, index)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{index}{index of interval}
}
\value{
numeric
}
\description{
Returns start time of interval in interval tier.
}
\examples{
tg <- tg.sample()
tg.getIntervalStartTime(tg, "phone", 5)
}

