% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getLabel}
\alias{tg.getLabel}
\title{tg.getLabel}
\usage{
tg.getLabel(tg, tierInd, index)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{index}{index of point or interval}
}
\value{
character string
}
\description{
Return label of point or interval at the specified index.
}
\examples{
tg <- tg.sample()
tg.getLabel(tg, "phoneme", 4)
tg.getLabel(tg, "phone", 4)
}
\seealso{
\code{\link{tg.setLabel}}, \code{\link{tg.countLabels}}
}

