% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg.R
\name{tg.getNumberOfIntervals}
\alias{tg.getNumberOfIntervals}
\title{tg.getNumberOfIntervals}
\usage{
tg.getNumberOfIntervals(tg, tierInd)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}
}
\value{
integer
}
\description{
Returns number of intervals in the given interval tier.
}
\examples{
tg <- tg.sample()
tg.getNumberOfIntervals(tg, "phone")
}
\seealso{
\code{\link{tg.getNumberOfPoints}}
}
