% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getNumberOfPoints}
\alias{tg.getNumberOfPoints}
\title{tg.getNumberOfPoints}
\usage{
tg.getNumberOfPoints(tg, tierInd)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}
}
\value{
integer
}
\description{
Returns number of points in the given point tier.
}
\examples{
tg <- tg.sample()
tg.getNumberOfPoints(tg, "phoneme")
}

