% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getPointIndexLowerThanTime}
\alias{tg.getPointIndexLowerThanTime}
\title{tg.getPointIndexLowerThanTime}
\usage{
tg.getPointIndexLowerThanTime(tg, tierInd, time)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{time}{time which is going to be found in points}
}
\value{
integer
}
\description{
Returns index of point which is nearest the given time from left, i.e.
pointTime <= time. Tier index must belong to point tier.
}
\examples{
tg <- tg.sample()
tg.getPointIndexLowerThanTime(tg, "phoneme", 0.5)
}
\seealso{
\code{\link{tg.getPointIndexNearestTime}}, \code{\link{tg.getPointIndexHigherThanTime}}, \code{\link{tg.getLabel}}
}

