% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getPointIndexNearestTime}
\alias{tg.getPointIndexNearestTime}
\title{tg.getPointIndexNearestTime}
\usage{
tg.getPointIndexNearestTime(tg, tierInd, time)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{time}{time which is going to be found in points}
}
\value{
integer
}
\description{
Returns index of point which is nearest the given time (from both sides).
Tier index must belong to point tier.
}
\examples{
tg <- tg.sample()
tg.getPointIndexNearestTime(tg, "phoneme", 0.5)
}
\seealso{
\code{\link{tg.getPointIndexLowerThanTime}}, \code{\link{tg.getPointIndexHigherThanTime}}, \code{\link{tg.getLabel}}
}

