% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getPointTime}
\alias{tg.getPointTime}
\title{tg.getPointTime}
\usage{
tg.getPointTime(tg, tierInd, index)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{index}{index of point}
}
\value{
numeric
}
\description{
Return time of point at the specified index in point tier.
}
\examples{
tg <- tg.sample()
tg.getPointTime(tg, "phoneme", 4)
}
\seealso{
\code{\link{tg.getLabel}}, \code{\link{tg.getPointIndexNearestTime}}, \code{\link{tg.getPointIndexLowerThanTime}},

\code{\link{tg.getPointIndexHigherThanTime}}
}
