% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getStartTime}
\alias{tg.getStartTime}
\title{tg.getStartTime}
\usage{
tg.getStartTime(tg, tierInd = 0)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{[optional] tier index or "name"}
}
\value{
numeric
}
\description{
Returns start time. If tier index is specified, it returns start time
of the tier, if it is not specified, it returns start time of the whole
TextGrid.
}
\examples{
tg <- tg.sample()
tg.getStartTime(tg)
tg.getStartTime(tg, "phone")
}

