% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getTierName}
\alias{tg.getTierName}
\title{tg.getTierName}
\usage{
tg.getTierName(tg, tierInd)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}
}
\value{
character string
}
\description{
Returns name of the tier.
}
\examples{
tg <- tg.sample()
tg.getTierName(tg, 2)
}
\seealso{
\code{\link{tg.setTierName}}, \code{\link{tg.isIntervalTier}}, \code{\link{tg.isPointTier}}
}
