% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getTotalDuration}
\alias{tg.getTotalDuration}
\title{tg.getTotalDuration}
\usage{
tg.getTotalDuration(tg, tierInd = 0)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{[optional] tier index or "name"}
}
\value{
numeric
}
\description{
Returns total duration. If tier index is specified, it returns duration
of the tier, if it is not specified, it returns total duration of the
TextGrid.
}
\examples{
tg <- tg.sample()
tg.getTotalDuration(tg)
tg.getTotalDuration(tg, "phone")
}
\seealso{
\code{\link{tg.getStartTime}}, \code{\link{tg.getEndTime}}
}
