% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg.R
\name{tg.insertPoint}
\alias{tg.insertPoint}
\title{tg.insertPoint}
\usage{
tg.insertPoint(tg, tierInd, time, label)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{time}{time of the new point}

\item{label}{time of the new point}
}
\value{
TextGrid object
}
\description{
Inserts new point to point tier of the given index.
}
\examples{
\dontrun{
tg <- tg.sample()
tg2 <- tg.insertPoint(tg, "phoneme", 1.4, "NEW POINT")
tg.plot(tg2)
}
}
\seealso{
\code{\link{tg.removePoint}}, \code{\link{tg.insertInterval}}, \code{\link{tg.insertBoundary}}
}
