% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.isPointTier}
\alias{tg.isPointTier}
\title{tg.isPointTier}
\usage{
tg.isPointTier(tg, tierInd)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}
}
\value{
TRUE / FALSE
}
\description{
Returns TRUE if the tier is PointTier, FALSE otherwise.
}
\examples{
tg <- tg.sample()
tg.isPointTier(tg, 1)
tg.isPointTier(tg, "word")
}
\seealso{
\code{\link{tg.isIntervalTier}}, \code{\link{tg.getTierName}}
}

