% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.removePoint}
\alias{tg.removePoint}
\title{tg.removePoint}
\usage{
tg.removePoint(tg, tierInd, index)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{index}{index of point to be removed}
}
\value{
TextGrid object
}
\description{
Remove point of the given index from the point tier.
}
\examples{
tg <- tg.sample()
tg$phoneme$label
tg2 <- tg.removePoint(tg, "phoneme", 1)
tg2$phoneme$label
}
\seealso{
\code{\link{tg.insertPoint}}, \code{\link{tg.getNumberOfPoints}}, \code{\link{tg.removeIntervalBothBoundaries}}
}

