% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.removeTier}
\alias{tg.removeTier}
\title{tg.removeTier}
\usage{
tg.removeTier(tg, tierInd)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}
}
\value{
TextGrid object
}
\description{
Removes tier of the given index.
}
\examples{
\dontrun{
tg <- tg.sample()
tg.plot(tg)
tg2 <- tg.removeTier(tg, "word")
tg.plot(tg2)
}
}
\seealso{
\code{\link{tg.insertNewIntervalTier}}, \code{\link{tg.insertNewPointTier}}, \code{\link{tg.duplicateTier}}
}

