% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat_sampleData.R
\name{tg.sample}
\alias{tg.sample}
\title{tg.sample}
\usage{
tg.sample()
}
\value{
TextGrid
}
\description{
Returns sample TextGrid.
}
\examples{
tg <- tg.sample()
tg.plot(tg)
}
\seealso{
\code{\link{tg.plot}}
}

