% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat_sampleData.R
\name{tg.sampleProblem}
\alias{tg.sampleProblem}
\title{tg.sampleProblem}
\usage{
tg.sampleProblem()
}
\value{
TextGrid
}
\description{
Returns sample TextGrid with continuity problem.
}
\examples{
tg <- tg.sampleProblem()
tg2 <- tg.repairContinuity(tg)
tg2 <- tg.repairContinuity(tg2)
tg.plot(tg2)
}
\seealso{
\code{\link{tg.repairContinuity}}
}
