% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.setLabel}
\alias{tg.setLabel}
\title{tg.setLabel}
\usage{
tg.setLabel(tg, tierInd, index, newLabel)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{index}{index of interval or point}

\item{newLabel}{new "label"}
}
\description{
Sets (changes) label of interval or point of the given index in the
interval or point tier.
}
\examples{
tg <- tg.sample()
tg2 <- tg.setLabel(tg, "word", 3, "New Label")
tg.getLabel(tg2, "word", 3)
}

