% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg.R
\name{tg.setTierName}
\alias{tg.setTierName}
\title{tg.setTierName}
\usage{
tg.setTierName(tg, tierInd, name)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{tier index or "name"}

\item{name}{new "name" of the tier}
}
\description{
Sets (changes) name of tier of the given index.
}
\examples{
tg <- tg.sample()
tg2 <- tg.setTierName(tg, "word", "WORDTIER")
tg.getTierName(tg2, 4)
}
\seealso{
\code{\link{tg.getTierName}}
}
