% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tgCheckTierInd}
\alias{tgCheckTierInd}
\title{tgCheckTierInd}
\usage{
tgCheckTierInd(tg, tierInd)
}
\arguments{
\item{tg}{TextGrid object}

\item{tierInd}{Tier index or "name"}
}
\value{
Tier index
}
\description{
Returns tier index. Input can be either index (number) or tier name (character string).
It performs checks whether the tier exists.
}
\examples{
tg <- tg.sample()
tgCheckTierInd(tg, 4)
tgCheckTierInd(tg, 'word')
}

