% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{wrLine}
\alias{wrLine}
\title{wrLine}
\usage{
wrLine(string, fid)
}
\arguments{
\item{string}{Text line.}

\item{fid}{A connection object.}
}
\value{
a raw vector (if fid is a raw vector) or invisibly NULL.
}
\description{
Write text line to a connection in binary mode.
}
