% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it.R
\name{it.plot}
\alias{it.plot}
\title{it.plot}
\usage{
it.plot(it, group = "", snd = NULL)
}
\arguments{
\item{it}{IntensityTier object}

\item{group}{[optional] character string, name of group for dygraphs synchronization}

\item{snd}{[optional] Sound object}
}
\description{
Plots interactive IntensityTier using \code{dygraphs} package.
}
\examples{
\dontrun{
it <- it.sample()
it.plot(it)
}
}
\seealso{
\code{\link{it.read}}, \code{\link{tg.plot}}, \code{\link{it.cut}}, \code{\link{it.cut0}}, \code{\link{it.interpolate}}, \code{\link{it.write}}
}
