% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{it.write}
\alias{it.write}
\title{it.write}
\usage{
it.write(it, fileNameIntensityTier, format = "short")
}
\arguments{
\item{it}{IntensityTier object}

\item{fileNameIntensityTier}{file name to be created}

\item{format}{Output file format ("short" (short text format - default), "text" (a.k.a. full text format))}
}
\description{
Saves IntensityTier to file (in UTF-8 encoding).
it is list with at least $t and $i vectors (of the same length).
If there are no $tmin and $tmax values, there are
set as min and max of $t vector.
}
\examples{
\dontrun{
it <- it.sample()
it.plot(pt)
it.write(it, "demo/intensity.IntensityTier")
}
}
\seealso{
\code{\link{it.read}}, \code{\link{tg.write}}, \code{\link{it.interpolate}}
}
