% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{pt.legendre}
\alias{pt.legendre}
\title{pt.legendre}
\usage{
pt.legendre(pt, npoints = 1000)
}
\arguments{
\item{pt}{PitchTier object}

\item{npoints}{Number of points of PitchTier interpolation}
}
\value{
Vector of Legendre polynomials
}
\description{
Interpolate the PitchTier in 'npoints' equidistant points and approximate it by Legendre polynomials
}
\examples{
pt <- pt.sample()
pt$f <- pt$f[pt$t >= 3]  # Cut the PitchTier - from time 3 seconds to the end
pt$t <- pt$t[pt$t >= 3] - 3
pt$tmax <- pt$tmax - 3
c <- pt.legendre(pt)
print(c)
leg <- pt.legendreSynth(c)
\dontrun{
pt.plot(pt)
plot(leg)
}
}
\seealso{
\code{\link{pt.legendreSynth}}, \code{\link{pt.legendreDemo}}, \code{\link{pt.read}}, \code{\link{pt.plot}}, \code{\link{pt.Hz2ST}}, \code{\link{pt.interpolate}}
}
