% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.getNumberOfTiers}
\alias{tg.getNumberOfTiers}
\title{tg.getNumberOfTiers}
\usage{
tg.getNumberOfTiers(tg)
}
\arguments{
\item{tg}{TextGrid object}
}
\value{
integer
}
\description{
Returns number of tiers.
}
\examples{
tg <- tg.sample()
tg.getNumberOfTiers(tg)
}

