% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpraat.R
\name{tg.plot}
\alias{tg.plot}
\title{tg.plot}
\usage{
tg.plot(tg, group = "", pt = NULL, it = NULL)
}
\arguments{
\item{tg}{TextGrid object}

\item{group}{[optional] character string, name of group for dygraphs synchronization}

\item{pt}{[optional] PitchTier object}

\item{it}{[optional] IntensityTier object}
}
\description{
Plots interactive TextGrid using dygraphs package.
}
\examples{
\dontrun{
tg <- tg.sample()
tg.plot(tg)
tg.plot(tg.sample(), pt = pt.sample())
}
}
\seealso{
\code{\link{tg.read}}, \code{\link{pt.plot}}
}
